Refactored prompt type to lib
